+++ src/index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
     <canvas id="gameCanvas" width="800" height="600"></canvas>
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
